+++ pychron/core/progress.py
@@ -24,12 +24,12 @@ class CancelLoadingError(BaseException):
     pass
 
 
-def open_progress(n, close_at_end=True):
+def open_progress(n, close_at_end=True, **kw):
     pd = myProgressDialog(max=n - 1,
                           #dialog_size=(0,0, 550, 15),
                           close_at_end=close_at_end,
                           can_cancel=True,
-                          can_ok=True)
+                          can_ok=True, **kw)
     pd.open()
     return pd